Add OCR parsing test and export parser utilities
